Retyped to float

+++ SpaceInvaders/EntityRelated/Coordinate2D.cs
@@ -8,16 +8,16 @@ namespace SpaceInvaders.EntitiyRelated
 {
     internal class Coordinate2D
     {
-        private int x;
-        private int y;
+        private float x;
+        private float y;
 
-        public Coordinate2D(int x, int y)
+        public Coordinate2D(float x, float y)
         {
             X = x;
             Y = y;
         }
 
-        public int X { get => x; set => x = value; }
-        public int Y { get => y; set => y = value; }
+        public float X { get => x; set => x = value; }
+        public float Y { get => y; set => y = value; }
     }
 }
